ai-service/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# OCR
pytesseract==0.3.13
Pillow==11.0.0
pdf2image==1.17.0

# PDF Generation
reportlab==4.2.5
PyPDF2==3.0.1

# AI / OpenAI
openai==1.55.0

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36

# Utils
python-dotenv==1.0.1
pydantic==2.10.0
pydantic-settings==2.6.1
httpx==0.28.0
aiofiles==24.1.0

# Document processing
python-docx==1.1.2
openpyxl==3.1.5
